fix: text refactoring
